refactor(interactivity): extract resolveHyperlinkHandling from domain-utils
Move link-click delegation into src/interactivity/hyperlinks.ts and drop
the now-orphaned IVisualHost import from domain-utils. The interactivity
surface is now consolidated. No behaviour change.
