+++ src/interactivity/index.ts
@@ -2,3 +2,4 @@
 // rather than reaching into individual modules.
 export { BehaviorManager, IHtmlBehaviorOptions } from './behavior';
 export { resolveHover } from './tooltips';
+export { resolveHyperlinkHandling } from './hyperlinks';
